apps/risk-evidence-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
prometheus_client==0.20.0
pyyaml==6.0.2
